backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
langchain==0.0.350
langchain-google-genai==0.0.5
faiss-cpu>=1.8.0
python-dotenv==1.0.0